+++ src/services/index.js
@@ -9,3 +9,5 @@ exports["xterm-terminal"] = require("./xterm-terminal.js");
 exports["ftp-service"] = require("./ftp-service.js");
 exports["config-service"] = require("./config-service.js");
 exports["credential-service"] = require("./credential-service.js");
+exports["ssh-credential-service"] = require("./ssh-credential-service.js");
+exports["ftp-credential-service"] = require("./ftp-credential-service.js");